ExtractData.py
Our purpose is to extract information from a given .pdf file. I do it with PDFMiner3k, which is compatible with Python3.6. It can extract all the text and output it as a string. But in this moment, we haven't integrated these code into our project yet because there's still some problems to be solved